% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_data_sim}
\alias{main_data_sim}
\title{Simulated Main-Study Data}
\format{
A data frame with 89,538 observations and the following variables:
\describe{
\item{id}{Participant identifier.}
\item{fqcal}{Food-frequency questionnaire estimate of caloric intake.}
\item{fqcalinc}{Transformed caloric intake.}
\item{fqtfat}{Food-frequency questionnaire estimate of total fat intake.}
\item{fqtfatinc}{Transformed total fat intake.}
\item{fqalc}{Food-frequency questionnaire estimate of alcohol intake.}
\item{fqalcinc}{Transformed alcohol intake.}
\item{age}{Participant age.}
\item{agec}{Categorized participant age.}
\item{case}{Binary outcome indicator.}
}
}
\source{
Simulated data created for the RegCalib package.
}
\usage{
main_data_sim
}
\description{
A simulated main-study dataset for demonstrating regression calibration
with error-prone dietary exposures.
}
\keyword{datasets}
