% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_data_sim}
\alias{valid_data_sim}
\title{Simulated External Validation Data}
\format{
A data frame with 173 observations and variables containing
food-frequency questionnaire measurements, reference measurements, and age.
\describe{
\item{id}{Participant identifier.}
\item{fqcal}{Surrogate caloric intake measurement.}
\item{fqcalinc}{Transformed surrogate caloric intake.}
\item{fqtfat}{Surrogate total-fat intake measurement.}
\item{fqtfatinc}{Transformed surrogate total-fat intake.}
\item{fqalc}{Surrogate alcohol intake measurement.}
\item{fqalcinc}{Transformed surrogate alcohol intake.}
\item{drcal}{Reference caloric intake measurement.}
\item{drcalinc}{Transformed reference caloric intake.}
\item{drtfat}{Reference total-fat intake measurement.}
\item{drtfatinc}{Transformed reference total-fat intake.}
\item{dralc}{Reference alcohol intake measurement.}
\item{dralcinc}{Transformed reference alcohol intake.}
\item{age}{Participant age.}
\item{agec}{Categorized participant age.}
}
}
\source{
Simulated data created for the RegCalib package.
}
\usage{
valid_data_sim
}
\description{
A simulated external validation dataset containing surrogate and reference
measurements of dietary exposures.
}
\keyword{datasets}
